fix: update zigzagDecode return type to int and adjust tests for consistency

+++ clients/dart/lib/lipi/zigzag.dart
@@ -2,6 +2,6 @@ BigInt zigzagEncode(BigInt n) {
   return (n << 1) ^ (n >> 63);
 }
 
-BigInt zigzagDecode(BigInt n) {
-  return (n >> 1) ^ -(n & BigInt.one);
+int zigzagDecode(BigInt n) {
+  return ((n >> 1) ^ -(n & BigInt.one)).toInt();
 }
